database_manager/requirements.txt
FLASK
FLASK-SQLALCHEMY
FLASK-LOGIN-0.4.1